refactor: Move component tests to its correct folder
