+++ src/components/__tests__/RestaurantDetails.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen, waitFor } from "@testing-library/react";
-import RestaurantList from "../components/RestaurantList";
+import RestaurantDetails from "../RestaurantDetails";
 
 beforeEach(() => {
   global.fetch = jest.fn();
@@ -9,7 +9,12 @@ afterEach(() => {
   jest.restoreAllMocks();
 });
 
-describe("<RestaurantList />", () => {
+describe("<RestaurantDetails />", () => {
+  it("renders empty if not restaurantId is invalid", () => {
+    renderComponent({ restaurantId: 0 });
+
+    expect(screen.queryByText(/Restaurant Details/)).toBeNull();
+  });
   it("handles error from server", async () => {
     // Given: A server error
     (global.fetch as jest.Mock).mockResolvedValueOnce({
@@ -17,17 +22,17 @@ describe("<RestaurantList />", () => {
     });
 
     // When: Render a component
-    renderComponent();
+    renderComponent({ restaurantId: 1 });
 
     // Then: The component renders an error message
     await waitFor(() => {
       expect(
-        screen.getByText(/Error fetching restaurants: Server error/)
+        screen.getByText(/Error fetching restaurant details: Server error/)
       ).toBeInTheDocument();
     });
   });
 });
 
-function renderComponent() {
-  return render(<RestaurantList onRestaurantSelect={jest.fn()} />);
+function renderComponent({ restaurantId }: { restaurantId: number }) {
+  return render(<RestaurantDetails restaurantId={restaurantId} />);
 }
