+++ src/components/__tests__/RestaurantList.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen, waitFor } from "@testing-library/react";
-import RestaurantDetails from "../components/RestaurantDetails";
+import RestaurantList from "../RestaurantList";
 
 beforeEach(() => {
   global.fetch = jest.fn();
@@ -9,12 +9,7 @@ afterEach(() => {
   jest.restoreAllMocks();
 });
 
-describe("<RestaurantDetails />", () => {
-  it("renders empty if not restaurantId is invalid", () => {
-    renderComponent({ restaurantId: 0 });
-
-    expect(screen.queryByText(/Restaurant Details/)).toBeNull();
-  });
+describe("<RestaurantList />", () => {
   it("handles error from server", async () => {
     // Given: A server error
     (global.fetch as jest.Mock).mockResolvedValueOnce({
@@ -22,17 +17,17 @@ describe("<RestaurantDetails />", () => {
     });
 
     // When: Render a component
-    renderComponent({ restaurantId: 1 });
+    renderComponent();
 
     // Then: The component renders an error message
     await waitFor(() => {
       expect(
-        screen.getByText(/Error fetching restaurant details: Server error/)
+        screen.getByText(/Error fetching restaurants: Server error/)
       ).toBeInTheDocument();
     });
   });
 });
 
-function renderComponent({ restaurantId }: { restaurantId: number }) {
-  return render(<RestaurantDetails restaurantId={restaurantId} />);
+function renderComponent() {
+  return render(<RestaurantList onRestaurantSelect={jest.fn()} />);
 }
